requirements.txt
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.4.0
sympy>=1.8.0
numba>=0.54.0
pytest>=6.2.5
cupy-cuda11x>=10.0.0; platform_system != "Windows"  # 可选GPU加速